code cleanup and bug fixing

+++ src/main/java/pl/patrykkukula/MovieReviewPortal/Service/Impl/AuthServiceImpl.java
@@ -19,6 +19,7 @@ import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
+import static java.lang.System.*;
 
 @Service
 @RequiredArgsConstructor
@@ -31,7 +32,7 @@ public class AuthServiceImpl implements IAuthService {
 
     @Override
     @Transactional
-    public void register(UserEntityDto userDto) {
+    public String register(UserEntityDto userDto) {
         Optional<UserEntity> existingUser = userRepository.findUserByUsernameOrEmail(userDto.getUsername(), userDto.getEmail());
         if (existingUser.isPresent()) throw new IllegalStateException("Username or email already exists");
 
@@ -46,43 +47,56 @@ public class AuthServiceImpl implements IAuthService {
                         .build();
         userRepository.save(user);
 
-       generateVerificationToken(user);
+       return generateVerificationToken(user);
     }
     @Override
     @Transactional
     public void verifyAccount(String token) {
+        if (token == null || token.isEmpty()) throw new IllegalArgumentException("token cannot be null or empty");
         VerificationToken verificationToken = verificationRepository.findByToken(token)
                 .orElseThrow(() -> new IllegalStateException("Invalid or expired token"));
         if (verificationToken.getExpiryDate().isBefore(LocalDateTime.now())) throw new IllegalStateException("Token expired");
         UserEntity userEntity = verificationToken.getUser();
-
+        if (userEntity.isEnabled()) throw new IllegalStateException("Account already verified");
         userEntity.setEnabled(true);
         userRepository.save(userEntity);
         verificationRepository.delete(verificationToken);
     }
     @Override
     @Transactional
-    public void resendVerificationToken(String username) {
-        UserEntity user = userRepository.findByUsername(username).orElseThrow(() -> new ResourceNotFoundException("Username", "username", username));
-        generateVerificationToken(user);
+    public String resendVerificationToken(String email) {
+        if (email == null || email.isEmpty()) throw new IllegalArgumentException("email cannot be null or empty");
+        UserEntity user = userRepository.findByEmail(email).orElseThrow(() -> new ResourceNotFoundException("Account", "email", email));
+        return generateVerificationToken(user);
     }
     @Override
     @Transactional
     public void resetPassword(PasswordResetDto passwordResetDto) {
-        PasswordResetToken token = pwdRepository.findByTokenWithUser(passwordResetDto.getPwdResetToken())
+        PasswordResetToken token = pwdRepository.findByToken(passwordResetDto.getPwdResetToken())
                 .orElseThrow(() -> new IllegalStateException("Invalid token"));
         if (token.getExpiryDate().isBefore(LocalDateTime.now())) throw new IllegalStateException("Token expired");
-        UserEntity user = token.getUser();
-        String hashedPassword = encoder.encode(passwordResetDto.getPassword());
-        if (hashedPassword.equals(user.getPassword())) throw new IllegalStateException("New password cannot be same as old password");
-        user.setPassword(hashedPassword);
+        UserEntity user = userRepository.findByEmail(passwordResetDto.getEmail())
+                .orElseThrow(() -> new ResourceNotFoundException("Account", "email", passwordResetDto.getEmail()));
+
+        if (encoder.matches(passwordResetDto.getNewPassword(), user.getPassword())) throw new IllegalStateException("New password cannot be same as old password");
+        String hashedNewPassword = encoder.encode(passwordResetDto.getNewPassword());
+        user.setPassword(hashedNewPassword);
+        user.setPasswordResetToken(null);
         userRepository.save(user);
     }
     @Override
+    @Transactional
     public String generatePasswordResetToken(String email){
-        UserEntity user = userRepository.findByUsername(email)
+        UserEntity user = userRepository.findByEmail(email)
                 .orElseThrow(() -> new ResourceNotFoundException("User", "email", email));
-        pwdRepository.deleteByUser(user);
+        Optional<PasswordResetToken> existingToken = pwdRepository.findByUser(user);
+        existingToken.ifPresent(token -> {
+            user.setPasswordResetToken(null);
+            userRepository.save(user);
+            pwdRepository.delete(token);
+            pwdRepository.flush();
+        });
+
         String token = UUID.randomUUID().toString();
         PasswordResetToken pwdToken = PasswordResetToken.builder()
                 .token(token)
@@ -90,12 +104,19 @@ public class AuthServiceImpl implements IAuthService {
                 .user(user)
                 .build();
         pwdRepository.save(pwdToken);
-        return "Password reset token has been created and will be valid for 12 hours: " + token + System.lineSeparator() +
-        "Please input token and new password in form";
+        return "Password reset token has been created and will be valid for 12 hours: " + token + lineSeparator() +
+                "To reset your password, send a POST request to the following URL: " +
+                "http://localhost:8081/auth/reset with the following JSON body: " +
+                "{\"password\":" + "your new password" + lineSeparator() +
+                "\"token\": \"" + token + "\"" + "}";
     }
-    private void generateVerificationToken(UserEntity user){
+
+    private String generateVerificationToken(UserEntity user){
         if (user.isEnabled()) throw new IllegalStateException("Account is verified");
-        verificationRepository.deleteByUser(user);
+        user.setVerificationToken(null);
+        userRepository.save(user);
+        pwdRepository.flush();
+
         String token = UUID.randomUUID().toString();
         VerificationToken verificationToken = VerificationToken
                 .builder()
@@ -104,8 +125,9 @@ public class AuthServiceImpl implements IAuthService {
                 .user(user)
                 .build();
         verificationRepository.save(verificationToken);
-        String activationLink = "http://localhost:8080/register/confirm?token=" + token;
-        System.out.println("Verification token has been created and will be valid for 12 hours: ");
-        System.out.println("Click the link to activate account: " + activationLink);
+        return  "Verification token has been created and will be valid for 12 hours: " + lineSeparator() +
+                "To activate your account, send a POST request to the following URL: " +
+                "http://localhost:8081/auth/register/confirm with the following JSON body: " +
+                "{ \"token\": \"" + token + "\" }";
     }
 }
